+++ CHANGES.rst
@@ -101,6 +101,10 @@ Release History
 - ``LinearFilter`` now has the discretization method as a class attribute,
   instead of an argument to ``make_step``, to make it easier to change.
   (`#1535 <https://github.com/nengo/nengo/pull/1535>`__)
+- ``Synapse.filt`` no longer uses the initial value of the input signal to
+  initialize the state if ``y0 is None``. This allows unstable filters
+  (e.g. integrators) to be used with ``filt`` with the default of ``y0=0``.
+  (`#1535 <https://github.com/nengo/nengo/pull/1535>`__)
 
 **Deprecated**